+++ HISTORY.rst
@@ -7,6 +7,8 @@ Release History
 * Disable project auto-clean while building/uploading firmware using
   `platformio run --disable-auto-clean <http://docs.platformio.org/en/latest/userguide/cmd_run.html#cmdoption--disable-auto-clean>`_ option
   (`issue #255 <https://github.com/platformio/platformio/issues/255>`_)
+* Handle "OSError: [Errno 13] Permission denied" for PlatformIO installer script
+  (`issue #254 <https://github.com/platformio/platformio/issues/254>`_)
 
 2.2.1 (2015-07-17)
 ------------------